Add files via upload
This database contains files from an experiment on Kinesthetic Motor Imagery (KMI) levels. KMI is a robust paradigm that can be used for Brain-Computer Interface systems. KMI consists of imagining the sensations when a limb executes a task, such as squeezing a ball with a right hand. This task can be performed with low or high force. 

In this database, we acquired EEG signals from 10 healthy subjects. We asked to perform KMI at three levels of force (10%, 40$, and 70% of their Maximal Voluntary Contraction MVC) and the basal to have four classes. Each subject was labelled as user001, for example. They performed each level twice; you can identify it as user001_10_1.txt for user 1, performing 10% of their MVC in the first attempt. The file user004_70_2.txt belongs to user 4, who performed 70% of their MVC on the second attempt. 

The total time of the task is 40 seconds. The equipment that was used was the Emotiv Epoc, which works with a sampling time of 128 Hz. 

The KMI events were elicited at 2.9 sec, 10.9 sec, 18.9 sec, 26.9 sec, and 34.9 sec, making a total of five trials per file.